requirements.txt
Pillow==8.4.0
numpy==1.21.2
matplotlib==3.4.3
scipy==1.7.1
ultralytics==8.0.0
ruamel.yaml==0.17.17
scikit-learn==0.24.2
apache-airflow==2.5.0 